Cypher Keys: mechanical key sounds, multi-touch chording, SD sample packs
Project 21 grows from a working HID deck into something that feels like a real
keyboard, and four sibling projects land their own hardware proofs.

Project 21 — Cypher Keys HID Deck (field-proven on V1.2)
  * KeyAudio: mechanical key-click audio out of the NS4168 speaker. Three
    synthesized switch profiles (Blue/Brown/Red) rendered at boot as PCM, three
    pitch/level variants each so typing does not machine-gun, 4-voice mixing,
    fired at touch-DOWN where a real switch actuates. Reuses project 09's proven
    IDF i2s_std bring-up, including the silence-prefill before amp enable that
    avoids a startup pop, with the amp level read from the (corrected) active-LOW
    HardwareProfile flag rather than hardcoded.
  * KeySoundPacks: real recorded switch samples from SD. Per-row press clips plus
    dedicated BACKSPACE/ENTER/SPACE, with a fallback chain (class -> row generic
    -> R0 -> synthesized) so ragged packs still sound right. Selectable in
    settings alongside the built-in profiles; scripts/convert-key-sounds.sh
    transcodes a kbsim-style tree to the 16-bit PCM layout the engine plays.
    Samples are deliberately NOT vendored: their provenance is undocumented
    upstream, so the synthesized profiles stay the shipped default.
  * KeysTouch: 5-contact multi-touch replaces the single-point TouchInput. Real
    modifier chording (hold Cmd, tap C) alongside the old one-shot sticky taps,
    hold-to-repeat on Backspace and the arrows, and press ownership resolved at
    touch-DOWN so sliding a finger off a key cannot fire a different one.
  * KeysLayout.h: every coordinate and its hit-test in one header, adopting
    project 09's UiLayout pattern, so drawing and touch can no longer disagree.
    This retires the bug class that already bit us once (a status-bar button
    overlapping text because the two lived in different files).
  * Settings screen (sound profile, volume, brightness, theme, idle dim) with its
    own dispatch, freeing the crowded status bar; plus a boot splash and idle dim.
  * Host test harness (scripts/test-cypher-keys.sh) compiles the real keyboard
    and touch sources against a shim and asserts the chording/sticky/repeat rules,
    GT911 flicker handling, and pack fallback order against the actual converted
    WAVs. It earned its keep: it caught a silent __has_include failure where
    SD_MMC compiled green but never linked, and flagged a row-mapping change.

  Known limit, documented rather than papered over: the trackpad is USB-only.
  Any BLE mouse report panics the NimBLE/esp_hosted stack while keyboard and
  consumer control over the same link are solid. Report-rate coalescing, a fast
  connection interval, and buffer flow control all failed to prevent it, and the
  panic backtrace is unreachable over USB (P4 panics go to UART0), so BLE mouse
  output is disabled by design and the UI says so. Risk register row 21.

Sibling projects
  * 02 Cypher Vision Cam: CamAudio and ImageViewer; viewfinder aspect fix.
  * 09 Cypher Tune MPC: generated builtin kit, kit samples, loop library.
  * 10 LiteGo Touch Coach: playable 9x9 Go with a Monte-Carlo opponent.
  * 14 ADS-B Radar: AdsbFormat extraction and refactor.
  * 22 Cypher Boy: NES core via vendored nofrendo; gwenesis table regen.

Docs record proof states against evidence: risk register rows 17 and 19 close
(USB HID and BLE HID both host-proven), row 21 opens for the BLE mouse panic,
and the proof matrix reflects what was actually observed on hardware versus what
remains host-tested only.

+++ projects/14-adsb-flight-tracker-radar/src/AdsbTypes.h
@@ -33,7 +33,10 @@ struct AdsbSnapshot {
   Aircraft ac[kMaxContacts];
   uint8_t count;         // valid entries in ac[]
   uint16_t totalSeen;    // raw aircraft in the last fetch (before the cap)
-  uint8_t rangeRingKm;   // outer display ring (set by the dashboard)
+  uint16_t rangeRingKm;  // outer display ring (set by the dashboard). uint16 so
+                         // `range 300` reaches the scope intact - as a uint8 it
+                         // wrapped to 44 km and the disc silently disagreed
+                         // with the header pill.
   uint32_t generatedMs;  // millis() of the fetch that produced this data
   const char *source;    // "airplanes.live" / "adsb.fi" / "mock" / "no-wifi"
 };
